Move classes into suitable java packages
