+++ src/main/java/Duke.java
@@ -1,3 +1,8 @@
+import duke.command.Command;
+import duke.tasklist.TaskList;
+import duke.ui.Ui;
+import duke.parser.Parser;
+import duke.storage.Storage;
 import java.time.format.DateTimeParseException;
 import java.util.Scanner;
