--- src/main/java/UnmarkCommand.java
@@ -1,18 +1,0 @@
-public class UnmarkCommand extends Command{
-    private int index;
-    public UnmarkCommand(int index){
-        super("unmark");
-        this.index = index;
-    }
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        if(index < 0 || index >= tasks.size()){
-            System.out.println("Invalid index, please try again.");
-        } else {
-            Task task = tasks.unmark(index);
-            ui.taskUnmarkedMessage(task);
-            storage.save(tasks.list());
-        }
-    }
-}
