+++ src/main/java/duke/command/AddCommand.java
@@ -1,3 +1,13 @@
+package duke.command;
+
+import duke.storage.Storage;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.Todo;
+import duke.tasklist.TaskList;
+import duke.ui.Ui;
+
 import java.time.LocalDate;
 
 public class AddCommand extends Command {
