+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.storage.Storage;
+import duke.tasklist.TaskList;
+import duke.ui.Ui;
+
 public class Command {
     protected String keyword;
