+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,0 +1,26 @@
+package duke.command;
+
+import duke.storage.Storage;
+import duke.task.Task;
+import duke.tasklist.TaskList;
+import duke.ui.Ui;
+
+public class DeleteCommand extends Command{
+    private int index;
+    public DeleteCommand(int index){
+        super("delete");
+        this.index = index;
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        if(index < 0 || index >= tasks.size()){
+            System.out.println("Invalid index, please try again.");
+        } else {
+            Task task = tasks.remove(index);
+            ui.taskDeleteMessage(task, tasks.size());
+            storage.save(tasks.list());
+        }
+    }
+}
+
