+++ src/main/java/duke/storage/Storage.java
@@ -1,10 +1,18 @@
-import java.io.*;
+package duke.storage;
+
+import duke.task.Task;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.ObjectInputStream;
+import java.io.ObjectOutputStream;
 import java.util.ArrayList;
 import java.util.List;
 
 public class Storage {
     private final String FILE_PATH;
-    private final  File FILE;
+    private final File FILE;
 
     public Storage(String filePath){
         this.FILE_PATH = filePath;
