+++ src/main/java/duke/task/Deadline.java
@@ -1,17 +1,20 @@
+package duke.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task {
+public class Deadline extends Task {
     private LocalDate date;
 
-    public Event(String name, LocalDate date) {
+    public Deadline(String name, LocalDate date) {
         super(name);
         this.date =date;
     }
 
     @Override
     public String toString(){
-        return "[E] " + super.toString() + " (at: "
+        return "[D] " + super.toString() + " (by: "
                 + date.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ")";
     }
 }
+
