+++ src/main/java/duke/task/Event.java
@@ -1,18 +1,19 @@
+package duke.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
-public class Deadline extends Task {
+public class Event extends Task {
     private LocalDate date;
 
-    public Deadline(String name, LocalDate date) {
+    public Event(String name, LocalDate date) {
         super(name);
         this.date =date;
     }
 
     @Override
     public String toString(){
-        return "[D] " + super.toString() + " (by: "
+        return "[E] " + super.toString() + " (at: "
                 + date.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ")";
     }
 }
-
