+++ src/main/java/duke/tasklist/TaskList.java
@@ -1,3 +1,6 @@
+package duke.tasklist;
+
+import duke.task.Task;
 import java.util.ArrayList;
 import java.util.List;
